refactored/generalized pan into spoon to setup kitchen support
