+++ spec/mocks/spoon/return_0.sh
