+++ spec/mocks/spoon/return_1.sh
