+++ spec/mocks/spoon/return_2.sh
