+++ spec/mocks/spoon/return_3.sh
