+++ spec/mocks/spoon/return_7.sh
