+++ spec/mocks/spoon/return_8.sh
