+++ spec/mocks/spoon/return_9.sh
