+++ spec/mocks/spoon/version.sh
