+++ spec/pdi/spoon/param_spec.rb
@@ -9,7 +9,7 @@
 
 require 'spec_helper'
 
-describe PDI::Param do
+describe PDI::Spoon::Param do
   let(:key)   { 'k1' }
   let(:value) { 'v1' }
   let(:value_with_space) { 'v 1' }
